Changes int to size_t where needed

+++ Additional project/Мillionaire's garage/Мillionaire's garage/Plane.cpp
@@ -1,6 +1,6 @@
 #include "Plane.h"
 
-Plane::Plane(int maxPassengers, int maxSpeed, const String& name, int year, double wingspan, double maxAltitude, const String& color, const String& model
+Plane::Plane(size_t maxPassengers, size_t maxSpeed, const String& name, size_t year, double wingspan, double maxAltitude, const String& color, const String& model
     , double maxStorage, double maxAirTime) : AirVehicle(maxPassengers, maxSpeed, name, year, wingspan, maxAltitude, color, model)
 {
     this->maxStorage = maxStorage;
