+++ Additional project/Мillionaire's garage/Мillionaire's garage/Plane.h
@@ -9,7 +9,7 @@ private:
 	double maxAirTime;
 
 public:
-	Plane(int maxPassengers, int maxSpeed, const String& name, int year, double wingspan,
+	Plane(size_t maxPassengers, size_t maxSpeed, const String& name, size_t year, double wingspan,
 		double maxAltitude, const String& color, const String& model,  double maxStorage, double maxAirTime);
 
 	Vehicle* clone() const override;
